broke up config file into components
