+++ src/config/editor.rs
@@ -1,0 +1,443 @@
+use crate::cli::VERSION;
+use crate::editor::Editor;
+use crate::ui::Feedback;
+use kaolinite::Loc;
+use mlua::prelude::*;
+
+impl LuaUserData for Editor {
+    fn add_fields<'lua, F: LuaUserDataFields<'lua, Self>>(fields: &mut F) {
+        fields.add_field_method_get("pasting", |_, editor| Ok(editor.paste_flag));
+        fields.add_field_method_get("cursor", |_, editor| {
+            let loc = editor.doc().char_loc();
+            Ok(LuaLoc {
+                x: loc.x,
+                y: loc.y + 1,
+            })
+        });
+        fields.add_field_method_get("document_name", |_, editor| {
+            let name = editor.doc().file_name.clone();
+            Ok(name)
+        });
+        fields.add_field_method_get("document_length", |_, editor| {
+            let len = editor.doc().len_lines();
+            Ok(len)
+        });
+        fields.add_field_method_get("version", |_, _| Ok(VERSION));
+        fields.add_field_method_get("current_document_id", |_, editor| Ok(editor.ptr));
+        fields.add_field_method_get("document_count", |_, editor| Ok(editor.doc.len()));
+        // DEPRECIATED
+        fields.add_field_method_get("help_visible", |_, _| Ok(false));
+        fields.add_field_method_get("document_type", |_, editor| {
+            let ext = editor
+                .doc()
+                .file_name
+                .as_ref()
+                .and_then(|name| Some(name.split('.').last().unwrap_or("")))
+                .unwrap_or("");
+            let file_type = kaolinite::utils::filetype(ext);
+            Ok(file_type)
+        });
+    }
+
+    fn add_methods<'lua, M: LuaUserDataMethods<'lua, Self>>(methods: &mut M) {
+        // Reload the configuration file
+        methods.add_method_mut("reload_config", |lua, editor, ()| {
+            if editor
+                .load_config(editor.config_path.clone(), &lua)
+                .is_err()
+            {
+                editor.feedback = Feedback::Error("Failed to reload config".to_string());
+            }
+            Ok(())
+        });
+        // Display messages
+        methods.add_method_mut("display_error", |_, editor, message: String| {
+            editor.feedback = Feedback::Error(message);
+            Ok(())
+        });
+        methods.add_method_mut("display_warning", |_, editor, message: String| {
+            editor.feedback = Feedback::Warning(message);
+            Ok(())
+        });
+        methods.add_method_mut("display_info", |_, editor, message: String| {
+            editor.feedback = Feedback::Info(message);
+            Ok(())
+        });
+        // Prompt the user
+        methods.add_method_mut("prompt", |_, editor, question: String| {
+            Ok(editor
+                .prompt(question)
+                .unwrap_or_else(|_| "error".to_string()))
+        });
+        // Edit commands (relative)
+        methods.add_method_mut("insert", |_, editor, text: String| {
+            for ch in text.chars() {
+                if let Err(err) = editor.character(ch) {
+                    editor.feedback = Feedback::Error(err.to_string());
+                }
+            }
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("remove", |_, editor, ()| {
+            if let Err(err) = editor.backspace() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("insert_line", |_, editor, ()| {
+            if let Err(err) = editor.enter() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("remove_line", |_, editor, ()| {
+            if let Err(err) = editor.delete_line() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            update_highlighter(editor);
+            Ok(())
+        });
+        // Cursor moving
+        methods.add_method_mut("move_to", |_, editor, (x, y): (usize, usize)| {
+            let y = y.saturating_sub(1);
+            editor.doc_mut().move_to(&Loc { x, y });
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_up", |_, editor, ()| {
+            editor.up();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_down", |_, editor, ()| {
+            editor.down();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_left", |_, editor, ()| {
+            editor.left();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_right", |_, editor, ()| {
+            editor.right();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("select_up", |_, editor, ()| {
+            editor.select_up();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("select_down", |_, editor, ()| {
+            editor.select_down();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("select_left", |_, editor, ()| {
+            editor.select_left();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("select_right", |_, editor, ()| {
+            editor.select_right();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("select_all", |_, editor, ()| {
+            editor.select_all();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("cut", |_, editor, ()| {
+            if let Err(err) = editor.cut() {
+                editor.feedback = Feedback::Error(err.to_string());
+            } else {
+                editor.feedback = Feedback::Info("Text cut to clipboard".to_owned());
+            }
+            Ok(())
+        });
+        methods.add_method_mut("copy", |_, editor, ()| {
+            if let Err(err) = editor.copy() {
+                editor.feedback = Feedback::Error(err.to_string());
+            } else {
+                editor.feedback = Feedback::Info("Text copied to clipboard".to_owned());
+            }
+            Ok(())
+        });
+        methods.add_method_mut("move_home", |_, editor, ()| {
+            editor.doc_mut().move_home();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_end", |_, editor, ()| {
+            editor.doc_mut().move_end();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_page_up", |_, editor, ()| {
+            editor.doc_mut().move_page_up();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_page_down", |_, editor, ()| {
+            editor.doc_mut().move_page_down();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_top", |_, editor, ()| {
+            editor.doc_mut().move_top();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_bottom", |_, editor, ()| {
+            editor.doc_mut().move_bottom();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_previous_word", |_, editor, ()| {
+            editor.prev_word();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("move_next_word", |_, editor, ()| {
+            editor.next_word();
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut(
+            "insert_at",
+            |_, editor, (text, x, y): (String, usize, usize)| {
+                let y = y.saturating_sub(1);
+                let location = editor.doc_mut().char_loc();
+                editor.doc_mut().move_to(&Loc { x, y });
+                for ch in text.chars() {
+                    if let Err(err) = editor.character(ch) {
+                        editor.feedback = Feedback::Error(err.to_string());
+                    }
+                }
+                editor.doc_mut().move_to(&location);
+                update_highlighter(editor);
+                Ok(())
+            },
+        );
+        methods.add_method_mut("remove_at", |_, editor, (x, y): (usize, usize)| {
+            let y = y.saturating_sub(1);
+            let location = editor.doc_mut().char_loc();
+            editor.doc_mut().move_to(&Loc { x, y });
+            if let Err(err) = editor.delete() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            editor.doc_mut().move_to(&location);
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("insert_line_at", |_, editor, (text, y): (String, usize)| {
+            let y = y.saturating_sub(1);
+            let location = editor.doc_mut().char_loc();
+            if y < editor.doc().len_lines() {
+                editor.doc_mut().move_to_y(y);
+                editor.doc_mut().move_home();
+                if let Err(err) = editor.enter() {
+                    editor.feedback = Feedback::Error(err.to_string());
+                }
+                editor.up();
+            } else {
+                editor.doc_mut().move_bottom();
+                if let Err(err) = editor.enter() {
+                    editor.feedback = Feedback::Error(err.to_string());
+                }
+            }
+            for ch in text.chars() {
+                if let Err(err) = editor.character(ch) {
+                    editor.feedback = Feedback::Error(err.to_string());
+                }
+            }
+            editor.doc_mut().move_to(&location);
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("remove_line_at", |_, editor, y: usize| {
+            let y = y.saturating_sub(1);
+            let location = editor.doc_mut().char_loc();
+            editor.doc_mut().move_to_y(y);
+            if let Err(err) = editor.delete_line() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            editor.doc_mut().move_to(&location);
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("open_command_line", |_, editor, ()| {
+            match editor.prompt("Command") {
+                Ok(command) => {
+                    editor.command = Some(command);
+                }
+                Err(err) => {
+                    editor.feedback = Feedback::Error(err.to_string());
+                }
+            }
+            Ok(())
+        });
+        methods.add_method_mut("previous_tab", |_, editor, ()| {
+            editor.prev();
+            Ok(())
+        });
+        methods.add_method_mut("next_tab", |_, editor, ()| {
+            editor.next();
+            Ok(())
+        });
+        methods.add_method_mut("new", |_, editor, ()| {
+            if let Err(err) = editor.new_document() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            Ok(())
+        });
+        methods.add_method_mut("open", |_, editor, ()| {
+            if let Err(err) = editor.open_document() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            Ok(())
+        });
+        methods.add_method_mut("save", |_, editor, ()| {
+            if let Err(err) = editor.save() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            Ok(())
+        });
+        methods.add_method_mut("save_as", |_, editor, ()| {
+            if let Err(err) = editor.save_as() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            Ok(())
+        });
+        methods.add_method_mut("save_all", |_, editor, ()| {
+            if let Err(err) = editor.save_all() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            Ok(())
+        });
+        methods.add_method_mut("quit", |_, editor, ()| {
+            if let Err(err) = editor.quit() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            Ok(())
+        });
+        methods.add_method_mut("undo", |_, editor, ()| {
+            if let Err(err) = editor.undo() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("redo", |_, editor, ()| {
+            if let Err(err) = editor.redo() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("search", |_, editor, ()| {
+            if let Err(err) = editor.search() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method_mut("replace", |_, editor, ()| {
+            if let Err(err) = editor.replace() {
+                editor.feedback = Feedback::Error(err.to_string());
+            }
+            update_highlighter(editor);
+            Ok(())
+        });
+        methods.add_method("get_character", |_, editor, ()| {
+            let loc = editor.doc().char_loc();
+            let ch = editor
+                .doc()
+                .line(loc.y)
+                .unwrap_or_else(|| "".to_string())
+                .chars()
+                .nth(loc.x)
+                .and_then(|ch| Some(ch.to_string()))
+                .unwrap_or_else(|| "".to_string());
+            Ok(ch)
+        });
+        methods.add_method_mut("get_character_at", |_, editor, (x, y): (usize, usize)| {
+            editor.doc_mut().load_to(y);
+            let y = y.saturating_sub(1);
+            let ch = editor
+                .doc()
+                .line(y)
+                .unwrap_or_else(|| "".to_string())
+                .chars()
+                .nth(x)
+                .and_then(|ch| Some(ch.to_string()))
+                .unwrap_or_else(|| "".to_string());
+            update_highlighter(editor);
+            Ok(ch)
+        });
+        methods.add_method("get_line", |_, editor, ()| {
+            let loc = editor.doc().char_loc();
+            let line = editor.doc().line(loc.y).unwrap_or_else(|| "".to_string());
+            Ok(line)
+        });
+        methods.add_method_mut("get_line_at", |_, editor, y: usize| {
+            editor.doc_mut().load_to(y);
+            let y = y.saturating_sub(1);
+            let line = editor.doc().line(y).unwrap_or_else(|| "".to_string());
+            update_highlighter(editor);
+            Ok(line)
+        });
+        methods.add_method_mut("move_to_document", |_, editor, id: usize| {
+            editor.ptr = id;
+            Ok(())
+        });
+        methods.add_method_mut("move_previous_match", |_, editor, query: String| {
+            editor.prev_match(&query);
+            update_highlighter(editor);
+            Ok(())
+        });
+        // DEPRECIATED
+        methods.add_method_mut("hide_help_message", |_, _, ()| Ok(()));
+        // DEPRECIATED
+        methods.add_method_mut("show_help_message", |_, _, ()| Ok(()));
+        methods.add_method_mut("set_read_only", |_, editor, status: bool| {
+            editor.doc_mut().read_only = status;
+            Ok(())
+        });
+        methods.add_method_mut("set_file_type", |_, editor, ext: String| {
+            let mut highlighter = editor
+                .config
+                .syntax_highlighting
+                .borrow()
+                .get_highlighter(&ext);
+            highlighter.run(&editor.doc().lines);
+            editor.highlighter[editor.ptr] = highlighter;
+            Ok(())
+        });
+    }
+}
+
+pub struct LuaLoc {
+    x: usize,
+    y: usize,
+}
+
+impl IntoLua<'_> for LuaLoc {
+    fn into_lua(self, lua: &Lua) -> std::result::Result<LuaValue<'_>, LuaError> {
+        let table = lua.create_table()?;
+        table.set("x", self.x)?;
+        table.set("y", self.y)?;
+        Ok(LuaValue::Table(table))
+    }
+}
+
+fn update_highlighter(editor: &mut Editor) {
+    if let Err(err) = editor.update_highlighter() {
+        editor.feedback = Feedback::Error(err.to_string());
+    }
+}
